requirements.txt
python-telegram-bot
ollama
pillow
python-dotenv
discogs-client
